// server/client/models/fetchUsersControllerHandle200UsersItem.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * IFAL Arapiraca Backend API
 * API para gerenciamento de problemas de infraestrutura do IFAL Arapiraca
 * OpenAPI spec version: 1.0.0
 */
import type { FetchUsersControllerHandle200UsersItemRole } from "./fetchUsersControllerHandle200UsersItemRole";

export type FetchUsersControllerHandle200UsersItem = {
  id?: string;
  name?: string;
  email?: string;
  position?: string;
  role?: FetchUsersControllerHandle200UsersItemRole;
};